Copy over runfiles library from Bazel
This uses `root_symlinks` to preserve the fixed runfiles path at which the library must be available.

Users will see an error if their runfiles trees contain both `@bazel_tools//tools/bash/runfiles` and `@rules_shell//shell/runfiles` with a Bazel version in which the former is not yet an alias to the latter. This ensures that conflicting versions can't be mixed.

In fact, strict runfiles collision checking will be applied to all targets depending on `@rules_shell//shell/runfiles`, which may break some builds. If it turns out to be too breaking, we could add private API to `ctx.runfiles` that allows to disable the strict check from Starlark.